refactor: rename label-required-for-button to button-label-required
